procfs: move internal declarations to private header
This separates the procfs API that is used by
device drivers from the API that is used
internally by the core kernel code. The next
step will be to move the definition of
proc_dir_entry and read_proc/write_proc, so we
have better control over new drivers attempting
to use them.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>