procfs: use i_private for driver data
Most device drivers do not need to know the internals
of proc_dir_entry, except for getting their own private
data back. This makes it possible to retrieve the
same data from inode->i_private, as long as it does
not get modified, and changes most users to reference
that instead of PDE(inode)->data.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>