staging/rtl8192x: remove debug level proc interface
The procfs interface used in both rtl8192e and rtl8192u
is archaic and also pointless. Replacing it with a named
and writable module parameter makes it possible to
instantly remove a fair amount of code without changing
any functionality.

Most importantly, it removes some of the remaining users
of the write_proc interface that are standing in the
way of cleaning up the proc file system.

The debug level should probably be replaced with per-netdev
settings as in other drivers, but that is a different
issue.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>