procfs: rename most proc_dir_entry members
Most parts of proc_dir_entry are not meant to be touched
by device drivers that merely want to add a file to /proc.

This reorders the members to have the commonly accessed
ones on the top, and renames the private ones to start
with pde_, which makes it easier to grep for existing
users.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>