procfs: add proc_remove helper
A number of drivers call the remove_proc_entry function with
the pde->name and pde->parent arguments, to remove an entry
they already know. Add a simple helper to remove an entry
by its pointer directly, to avoid dereferencing the pde
from drivers.

This lets us move the definition of proc_dir_entry to a
private header file later.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>