procfs: introduce proc_create_simple
None of the remaining users of create_proc_read_entry
does anything useful with the four middle arguments,
they all just write text into the page buffer, and
update the position with varying degrees of correctness.

Simply changing the prototype to only pass the page
buffer and the private data, and removing all the bogus
code around it saves 600 lines of code.

More importantly, it fixes a number of bugs in all those
files that handled short reads incorrectly, and it prevents
new drivers from making the same mistakes.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>