scsi/megaraid: remove semaphore usage
Counting semaphores are going away, so replace the usage
in the megaraid driver with straight wait queues.

The code in megaraid_mm.c becomes more readable by this,
as it is made clear what the driver is actually waiting
for.

In the sas portion of the driver, the semaphore is
replaced by an equivalent combination of atomic plus
wait_queue_head.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>