staging/comedi/vmk80xx: remove semaphore
The vmk80xx driver has evidently copied code from the
USB skeleton driver that limits the amount of memory
being allocated in the kernel for USB I/O.

This doesn't work in case of comedi drivers, where the
allocation takes place in the subsystem code, which
means the semaphore is completely pointless.

Since we are trying to get rid of counting semaphores
everywhere, remove this instance without a replacement.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>