staging/cxt1e1: replace semaphores with mutex/completion
The cxt1e1 driver has an OS abstraction layer to build
its own mutexes and completions on top of the Linux
semaphore. This is a bad idea, and since we are trying
to get rid of counting semaphores, replace the whole
abstraction code with direct calls to the Linux
completion and mutex code.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>