dm-raid1: convert counting semaphore to simple one
The dm-region-hash code contains logic to use a counting
semaphore to control the number of concurrent recovery
threads. However, the code has always just allowed a
single thread here, which leads to the conclusion that
the case with multiple threads is probably broken.

This simplifies the code to allow only one thread for
recovery. Further simplifications on top are possible.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>