[HACK] ARM: build kernels larger than 32MB
If the kernel binary size exceeds 32MB, we have to build with -mlong-calls,
but that isn't enough for the various direct calls from assembler, which
this tries to turn into "long" calling conventions as well, badly.

This patch breaks stuff at run-time, because I'm basically clueless
about ARM assembly, but can probably serve as a starting point for
someone who understands what they are doing, and it gets the kernel to
build.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>