ARM: normalize asm/io.h accessors
Platforms that have their own I/O accessors should provide the
complete API so we can build all drivers that use the more
obscure interfaces and build without warnings.

In case of ebsa110, this adds support for a few functions
that are actually broken on the hardware, turning build
errors into runtime-errors. This is normally not desired,
but the platform is so obsolete by now that this seems better
than sprinkling 'depends on !ARCH_EBSA110' across drivers
that are known to be broken, so we can get randconfig
working.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>