[SUPERCEDED] ARM: imx: imx6q_set_lpm is only defined for CONFIG_PM=y
apparently this is fixed already.

This ensures that we only call imx6q_set_lpm if CONFIG_PM
is enabled and we are building the pm-imx6q.c file.

Another fix that has been suggested for this is to always
build this file conditionally build only the parts of it
that are relevant only to CONFIG_PM.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Shawn Guo <shawn.guo@freescale.com>
Cc: Sascha Hauer <kernel@pengutronix.de>