video: exynos: allow S6E8AX0 to be a module
This driver depends on LCD_CLASS_DEVICE, which can be a
loadable module, but S6E8AX0 can only be built-in at
the moment, which can lead to link errors when the
class device code is not built-in.

Changing this to allow the driver as a module also
requires exporting the exynos_mipi_dsi_register_lcd_driver
symbol.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>