ARM: debug_ll: add nop implementation
When selecting DEBUG_LL_UART_NONE, we need an implementation to
include. For traditional platforms, this is mach/debug-macro.S,
but this is not always provided, as platforms are migrating to
generic implementations in include/debug/.

This adds a trivial debug/none.S implementation that gets used
when DEBUG_LL_UART_NONE is chosen.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>