[HACK] Qualcomm PMIC8XXX keypad never worked
Something in the MFD_PM8XXX driver must have changed recently that now allows
the KEYBOARD_PMIC8XXX driver to be enabled. Unfortunately, it does not build
because of a dependency on the <linux/mfd/pm8xxx/gpio.h> header file that
does not exist.

In order to keep 'make randconfig' working on my test box, I have disabled
this driver, but it would be nice of someone could have a look into fixing it.

Not for inclusion.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>