ARM: always define CONFIG_PHYS_OFFSET on !MMU
Starting with 006fa2599bf0 "ARM: fix noMMU kallsyms symbol filtering", we
use CONFIG_PHYS_OFFSET to calculate PAGE_OFFSET and have that passed
into kallsyms as the start address.

However, CONFIG_PHYS_OFFSET is defined only for platforms that don't
set NEED_MACH_MEMORY_H. It is possible to build a noMMU kernel for
these platforms by setting CONFIG_DRAM_BASE to the same value as the
PLAT_PHYS_OFFSET. This means we can in fact rely on DRAM_BASE to be
correct already, and implicitly define PHYS_OFFSET to the same value.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>