ARM: autoselect CPU_ARM7TDMI if needed
The atmel at91x40 platform is the only one using an ARM7TDMI,
and it should just always enable this core.

Once that is done, there is no need to offer the Kconfig option
to other platforms, and it can be turned into a silent symbol.
Similarly, the CPU_ARM9TDMI can become invisible now, but it
is actually not used anywhere at the moment.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>