ARM: avoid dynamic patching on BE32 kernels
The __patch_text function uses __opcode_to_mem_thumb16, which does
not have a valid definition for BE32 kernels. Nobody cares much
about BE32 any more, and I suspect this won't be fixed, but we
should disallow any configurations that require the code patching.

The two options that use the code in question are CONFIG_KPROBES
and CONFIG_JUMP_LABEL. Annotate them in Kconfig to ensure they
are never enabled when building for BE32.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>