ARM: do not select ARM_PATCH_PHYS_VIRT per platform
In an attempt to get platforms closer to multiplatform
support, this option was turned on for mach-integrator
and mach-shmobile-legacy. Unfortunately, this breaks
certain configurations that conflict with the dynamic
patching, such as XIP_KERNEL and no-MMU support.

The only reason to have this option unconditionally
enabled on multiplatform kernels is that it without it
we cannot boot from an arbitrary address, but there is
no reason to force it here, and we may want to improve
the logic for multiplatform in the future as well.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>