[REVISIT] ARM: disable UACCESS_WITH_MEMCPY and Kprobes for ARMv3
We cannot currently build a kernel with either of these
options when gcc thinks it is building for ARMv3. Just
disabling them helps us build more random configurations.

Note that we don't actually support ARMv3 at all any more,
but when building a kernel for RiscPC with a StrongARM
CPU, we have to pretend to run on ARMv3 anyway.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>