[REVISIT] ARM: disable UACCESS_WITH_MEMCPY and Kprobes for ARMv3
We cannot currently build a kernel with either of these
options when gcc thinks it is building for ARMv3. Just
disabling them helps us build more random configurations
avoiding:

/tmp/ccVYu6PG.s: Assembler messages:
/tmp/ccVYu6PG.s:21947: Error: selected processor does not support `strh r0,[r1,-r2]' in ARM mode
/tmp/ccVYu6PG.s:21984: Error: selected processor does not support `strheq r14,[r11,r12]' in ARM mode
/tmp/ccVYu6PG.s:22009: Error: selected processor does not support `strheq r14,[r13,r12]' in ARM mode
/tmp/ccVYu6PG.s:22034: Error: selected processor does not support `strheq r14,[r12,r13]' in ARM mode
/tmp/ccVYu6PG.s:22071: Error: selected processor does not support `strh r1,[r2,r3]!' in ARM mode
/tmp/ccVYu6PG.s:22108: Error: selected processor does not support `strhne r12,[r11,-r10]!' in ARM mode
/tmp/ccVYu6PG.s:22145: Error: selected processor does not support `strh r2,[r3],r4' in ARM mode

Note that we don't actually support ARMv3 at all any more,
but when building a kernel for RiscPC with a StrongARM
CPU, we have to pretend to run on ARMv3 anyway.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>