ARM: erratum 798181 only makes sense with MMU
The 798181 erratum workaround is defined in the smp_tlb code,
which does not get compiled when the MMU is turned off, but
the code still gets called from the generic setup code,
which leads to a link failure.

This changes the declaration for the errata workaround to
an empty function when we are building a NOMMU kernel.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>