[EXPERIMENTAL] ARM: sa1100/pxa: fix MTD_XIP build
On Wednesday 19 March 2014, Nicolas Pitre wrote:
> On Wed, 19 Mar 2014, Russell King - ARM Linux wrote:
> >
> > I don't think you can do that.  I believe xip_irqpending() has to be
> > inline so that the XIP code (located in RAM) can detect when an
> > interrupt is pending, suspend the MTD operation, switch the MTD back
> > to read mode, and then allow the kernel to run.
> >
> > I'd be nervous about this without Nicolas checking it, or it being
> > built and the resulting assembly inspected.
>
> At the very minimum, xip_irqpending() should be marked with the __xipram
> attribute.

In commit 3169663ac5902 "ARM: sa11x0/pxa: convert OS timer registers
to IOMEM", the definition of the OSCR macro was changed to be an
__iomem pointer, but the same register is also used by the XIP
code. This patch does the corresponding change here as well.

Since PXA now uses a local variable for the base address of the
ICIP register, the xip_irqpending function has to be moved
into irq.c in the process.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Russell King <linux@arm.linux.org.uk>