thermal: fix modular build
CONFIG_THERMAL is defined as a tristate Kconfig symbol, but
setting it to 'm' for module gives us bogus inline wrappers
for thermal_zone_of_sensor_register and thermal_zone_of_sensor_unregister
because of a broken #ifdef statement in linux/thermal.h.

This fixes the header file by using the IS_ENABLED() macro that
tests for both built-in and module versions of the symbol.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>