clocksource: efm32 require GENERIC_CLOCKEVENTS
The efm32 clocksource driver can be built on any platform
with CONFIG_COMPILE_TEST enabled, but the code depends on
GENERIC_CLOCKEVENTS and causes a build failure if that is
not set.

This adds GENERIC_CLOCKEVENTS as an additional Kconfig
dependency to ensure the driver always builds cleanly when
it is enabled.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>