[UNREPRODUCABLE] serial: pl011: fix build without SYSRQ
If CONFIG_MAGIC_SYSRQ is disabled, there is no 'sysrq' member
in struct uart_port, and we get a build error if we try to access
it.

This changes the pl011 driver to add a compile-time check on
the SUPPORT_SYSRQ macro, just like we do for a number of other
uart drivers already.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>