usb/gadget: s3c-hsotg needs the phy subsystem
This driver fails to build if CONFIG_GENERIC_PHY is disabled.
In order to make it build cleanly all the time, this patch
adds a 'select' statement for this Kconfig symbol.

The reason for not using 'depends on' here as we do in a lot of
other cases is consistency because all other drivers also use
'select' for GENERIC_PHY.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>