[OBSOLETE] usb: phy: isp1301-omap needs tps65010
The isp1301-omap driver cannot be built-in if the tps65010 driver
is a module, otherwise we get a link error from the reference to
the tps65010_set_vbus_draw function.

There is already a hack in the driver to work around the problem
of tps65010 being not available at all. This is an attempt
to avoid the problem by ensuring it always is, but that turned
out to be wrong, so don't apply.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Felipe Balbi <balbi@ti.com>