drm: exynos+bridge: fix modular build
The exynos DRM driver can be a loadable module, which means
the ptn3460_init() function it calls has to be exported.

Further, we have to prevent the exynos display driver from
being built-in if the ptn3460 driver is a module, that
would also lead to a link error. It's fine if both are
modules, or ptn3460 is built-in, and also if ptn3460 is
completely disabled, since it has inline wrappers that
stub out this driver in this case.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>