ARM: Kconfig: clean up platform selection
With an ARCH_MULTIPLATFORM kernel, the configuration menu is
rather inconsistent because platforms use various ways to show
their platform specific options. This is an attempt to clean
it up by making every platform that has its own options use
a 'menuconfig' statement that serves both to show a top-level
option in the upper platform menu, and as a way to group the
platform specific options.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>