ARM: Kconfig: move multiplatform CPU arch selection
The CPU architecture selection was initially meant as the starting point
for the platform selection in the same menu. This has turned out too
complicated, so nobody ever used it that way.

Instead, this moves the CPU architecture options into the same level
as the platform selection, and adds appropriate 'comment' statements
to group them better.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>