drm/shmobile: s/#if CONFIG_/#ifdef CONFIG_/
Using #if is not the correct way to test for a kernel Kconfig
macro, and results in a message like

drivers/gpu/drm/shmobile/shmob_drm_drv.c:300:5: warning: "CONFIG_PM_SLEEP" is not defined [-Wundef]

when the symbol is not defined. This patch changes it to
instead use #ifdef, as everyone else does.

Signed-off-by: Arnd Bergmann <arnd@arnd@arndb.de>
Cc: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Cc: David Airlie <airlied@linux.ie>
Cc: dri-devel@lists.freedesktop.org