drm/exynos: s/#if CONFIG_/#ifdef CONFIG_/
Using #if is not the correct way to test for a kernel Kconfig
macro, and results in a message like

drivers/gpu/drm/exynos/exynos_drm_dsi.c1472:5: warning: "CONFIG_PM_SLEEP" is not defined [-Wundef]

when the symbol is not defined. This patch changes it to
instead use #ifdef, as everyone else does.

Signed-off-by: Arnd Bergmann <arnd@arnd@arndb.de>
Cc: Inki Dae <inki.dae@samsung.com>
Cc: Joonyoung Shim <jy0922.shim@samsung.com>
Cc: Seung-Woo Kim <sw0312.kim@samsung.com>
Cc: Kyungmin Park <kyungmin.park@samsung.com>
Cc: David Airlie <airlied@linux.ie>
Cc: dri-devel@lists.freedesktop.org