[media] davinci: export vpif_lock
The vpif driver can be built as a loadable module, which causes the
vpif_display driver to be unloadable because the vpif_lock symbol
is not exported, as reported by Kbuild:

ERROR: "vpif_lock" [drivers/media/platform/davinci/vpif_display.ko] undefined!

This adds the missing EXPORT_SYMBOL_GPL statement.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: "Lad, Prabhakar" <prabhakar.csengg@gmail.com>
Cc: Sekhar Nori <nsekhar@ti.com>
Cc: Mauro Carvalho Chehab <m.chehab@samsung.com>