misc: atmel_pwm: only build for supported platforms
There is architecture code in mach-at91 that depends on the
CONFIG_ATMEL_PWM symbol in order to call the soc-specific
at91_add_device_pwm function. While all of this is about code
that will be removed in the future, using DT probing and
the PWM framework, we currently get a build failure:

arch/arm/mach-at91/built-in.o: In function `at91_pwm_leds':
arch/arm/mach-at91/leds.c:88: undefined reference to `at91_add_device_pwm'

This patch ensures we only try to build this driver on
platforms on which it will build and work.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>