net/wan: cosa needs virt_to_bus
The cosa driver is rather outdated and does not get built on most
platforms because it requires the ISA_DMA_API symbol. However
there are some ARM platforms that have ISA_DMA_API but no virt_to_bus,
and they get this build error when enabling the ltpc driver.

/git/arm-soc/drivers/net/wan/cosa.c: In function 'tx_interrupt':
/git/arm-soc/drivers/net/wan/cosa.c:1768:3: error: implicit declaration of function 'virt_to_bus' [-Werror=implicit-function-declaration]
   unsigned long addr = virt_to_bus(cosa->txbuf);
   ^

This adds another dependency in Kconfig to avoid that configuration.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>