rtc: sa1100: use platform resource for mmio
As a preparation for ARM MMP multiplatform support, this
changes the sa1100 rtc driver that is shared between sa1100,
pxa and mmp to no longer rely on platform header files for
its register definitions.

Instead it uses the memory resource that gets passed by the
static platform device creation, or from DT. Unfortunately,
the PXA platform device is lacking the memory resource at the
moment, so I'm leaving a fallback path that takes the
traditional register definition to calculate the address.

I do not expect we will ever see a kernel that supports
running on sa1100 and another platform with this driver,
so we don't need to get more fancy for the register offsets.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Russell King <linux@arm.linux.org.uk>
Cc: Eric Miao <eric.y.miao@gmail.com>
Cc: Haojian Zhuang <haojian.zhuang@gmail.com>
Cc: Alessandro Zummo <a.zummo@towertech.it>
Cc: rtc-linux@googlegroups.com