spi: pxa2xx: make MMP independent of header files
The pxa2xx spi driver supports a number of front-ends, including
one for PCI, but it does not support MMP at the moment, despite
a patch by Qiao Zhou to add pxa910 into the driver.

With a multiplatform kernel, it is possible to build the PCI
variant with CONFIG_ARCH_MMP enabled, which causes a build
failure because of the missing mach/dma.h header.

This changes the linux/spi/pxa2xx_spi.h header file to only
include mach/dma.h on CONFIG_ARCH_PXA builds, to allow
allyesconfig builds with MMP included.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Mark Brown <broonie@opensource.wolfsonmicro.com>
Cc: Qiao Zhou <zhouqiao@marvell.com>
Cc: Haojian Zhuang <haojian.zhuang@gmail.com>