mtd: nand: pxa3xx: disable DMA support on MMP
The MMP platform traditionally uses ad-hoc DMA engine support
by having the slave drivers write directly to the DMA register
set, just as we do for PXA.

Nowadays, MMP has a proper dmaengine driver, but that conflicts
with the old method, so we cannot enable both at the same time.

Independent of this, we want to have multiplatform support
on MMP, which means the mach/dma.h header file becomes inaccessible
to the nand driver.

This patch disables the ad-hoc DMA support in the pxa3xx
nand driver for MMP, leaving it enabled only for PXA,
to address both of the above issues.

A patch series from Daniel Mack exists that should address
the performance impact this has over the traditional method:
http://www.spinics.net/lists/arm-kernel/msg265162.html
This patch series is still under review and needs more
testing, but I hope we can merge it soon.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Eric Miao <eric.y.miao@gmail.com>
Cc: Haojian Zhuang <haojian.zhuang@gmail.com>
Cc: David Woodhouse <dwmw2@infradead.org>
Cc: Brian Norris <computersforpeace@gmail.com>
Cc: Daniel Mack <zonque@gmail.com>