ARM: pxa: pass register resource for RTC
The sa1100 rtc device driver is shared between sa1100, pxa and mmp,
and multiplatform support for mmp requires to clean up that driver to
some degree, having it rely on resource passing instead of hardcoded
register numbers.

While sa1100, mmp+atags and mmp+dt all already have the correct resource
listed, the atags based PXA support only lists a resource for the rtc-pxa
device, not the rtc-sa1100 device. This is actually the same hardware block,
but used with different registers, and most of the defconfigs use the
rtc-sa1100 driver.

This patch sets up the mmio resource for the rtc-sa1100 device.
We can merge this independent of the patch to change the driver,
since that has a fallback (and a warning) to deal with the resource
missing on PXA.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Eric Miao <eric.y.miao@gmail.com>
Cc: Russell King <linux@arm.linux.org.uk>
Cc: Haojian Zhuang <haojian.zhuang@gmail.com>