ARM: make xscale iwmmxt code multiplatform aware
In a multiplatform configuration, we may end up building a kernel
for both Marvell PJ1 and an ARMv4 CPU implementation. In that
case, the xscale-cp0 code is built with gcc -march=armv4{,t},
which results in a build error from the coprocessor instructions.

Since we know this code will only have to run on an actual xscale
processor, we can simply build the entire file for ARMv5TE.

Related to this, we need to handle the iWMMXT initialization sequence
differently during boot, to ensure we don't try to touch xscale
specific registers on other CPUs from the xscale_cp0_init initcall.
cpu_is_xscale() used to be hardcoded to '1' in any configuration
that enables any XScale-compatible core, but this breaks once
we can have a combined kernel with MMP1 and something else.

In this patch, I introduce two new macros, cpu_is_xscale()
and cpu_is_mohawk, so we can test for all three of them in
the iwmmxt initialization. The two existing users of
cpu_is_xscale() are modified accordingly, but slightly
change behavior for kernels that enable CPU_MOHAWK in combination
with CPU_XSCALE or CPU_XSC3. Previously, these would leave
clear PMD_BIT4 in the page tables, now they leave it untouched,
just like we always do for kernels that enable only CPU_MOHAWK.
Since the previous behavior was inconsistent, I assume it was
unintentional.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>