[REVIEW] ARM: always build xscale-cp0 for ARMv5TE
In a multiplatform configuration, we may end up building a kernel
for both Marvell PJ1 and an ARMv4 CPU implementation. In that
case, the xscale-cp0 code is built with gcc -march=armv4{,t},
which results in a build error from the coprocessor instructions.

Since we know this code will only have to run on an actual xscale
processor, we can simply build the entire file for ARMv5TE.

[FIXME: we actually call the late_initcall on all CPUs, so
this seems unsafe].

Signed-off-by: Arnd Bergmann <arnd@arndb.de>